front: move sortPathOperationalPoints() to utils
This will make the next commit more readable.
