+++ front/src/applications/operationalStudies/hooks/__tests__/useSimulationResults.spec.ts
@@ -1,12 +1,9 @@
 import { cleanup, waitFor } from '@testing-library/react';
 import { renderHookWithStore } from 'store/__tests__';
-import { v4 as uuidV4 } from 'uuid';
 import { beforeEach, afterEach, describe, expect, it, vi } from 'vitest';
 
 import { mockOsrdEditoastEndpoints } from 'common/api/__mocks__/osrdEditoastApi';
 import type {
-  CoreOperationalPointOnPath,
-  PathItem,
   PacedTrainException,
   PathfindingResult,
   PathProperties,
@@ -14,7 +11,7 @@ import type {
   SimulationResponse,
 } from 'common/api/osrdEditoastApi';
 
-import useSimulationResults, { sortPathOperationalPoints } from '../useSimulationResults';
+import useSimulationResults from '../useSimulationResults';
 
 const INFRA_ID = 12;
 const ELECTRICAL_PROFILE_SET_ID = 34;
@@ -495,52 +492,5 @@ describe.skip('useSimulationResults', () => {
         exceptionId: undefined,
       });
     });
-  });
-});
-
-describe('sortPathOperationalPoints', () => {
-  const makeOp = (id: string, position: number) => ({ id, position }) as CoreOperationalPointOnPath;
-
-  const makePathItem = (opId: string): PathItem => ({
-    id: uuidV4(),
-    location: {
-      type: 'operational_point_part_reference',
-      operational_point: { type: 'id', operational_point: opId },
-    },
-  });
-
-  it('should not modify a list already sorted by position', () => {
-    const ops = [makeOp('A', 0), makeOp('B', 100), makeOp('C', 200)];
-    const path = [makePathItem('A'), makePathItem('B'), makePathItem('C')];
-    const result = sortPathOperationalPoints(ops, path);
-    expect(result.map((op) => op.id)).toEqual(['A', 'B', 'C']);
-  });
-
-  it('should put the path origin first in the list when its position matches the one of other waypoints', () => {
-    const ops = [makeOp('A', 0), makeOp('C', 0), makeOp('B', 0), makeOp('D', 200)];
-    const path = [makePathItem('C'), makePathItem('A'), makePathItem('D')];
-    const result = sortPathOperationalPoints(ops, path);
-    expect(result.map((op) => op.id)).toEqual(['C', 'A', 'B', 'D']);
-  });
-
-  it('should sort two path items based on their index in the path if they are next to each other in ops when they are neither the origin or destination)', () => {
-    const ops = [
-      makeOp('A', 0),
-      makeOp('C', 200),
-      makeOp('D', 200),
-      makeOp('B', 200),
-      makeOp('E', 200),
-      makeOp('F', 300),
-    ];
-    const path = [makePathItem('A'), makePathItem('B'), makePathItem('D'), makePathItem('F')];
-    const result = sortPathOperationalPoints(ops, path);
-    expect(result.map((op) => op.id)).toEqual(['A', 'C', 'B', 'D', 'E', 'F']);
-  });
-
-  it('should put the path destination at the end of the list when its position matches the one of other waypoints', () => {
-    const ops = [makeOp('A', 0), makeOp('D', 100), makeOp('B', 100), makeOp('C', 100)];
-    const path = [makePathItem('A'), makePathItem('C'), makePathItem('D')];
-    const result = sortPathOperationalPoints(ops, path);
-    expect(result.map((op) => op.id)).toEqual(['A', 'B', 'C', 'D']);
   });
 });
