+++ front/src/applications/operationalStudies/hooks/useSimulationResults.ts
@@ -4,12 +4,7 @@ import { skipToken } from '@reduxjs/toolkit/query';
 import { useTranslation } from 'react-i18next';
 import { useSelector } from 'react-redux';
 
-import {
-  osrdEditoastApi,
-  type CoreOperationalPointOnPath,
-  type PathItem,
-  type TrainScheduleResponse,
-} from 'common/api/osrdEditoastApi';
+import { osrdEditoastApi, type TrainScheduleResponse } from 'common/api/osrdEditoastApi';
 import formatPowerRestrictionRangesWithHandled from 'modules/powerRestriction/helpers/formatPowerRestrictionRangesWithHandled';
 import {
   extractOccurrenceDetailsFromPacedTrain,
@@ -28,37 +23,8 @@ import {
 } from 'utils/trainId';
 
 import type { SimulationResults } from '../types';
-import { matchOpRefAndWaypoint, preparePathPropertiesData } from '../utils';
+import { sortPathOperationalPoints, preparePathPropertiesData } from '../utils';
 import { useScenarioContext } from './useScenarioContext';
-
-/**
- * Sort operational points by position on the path (mm from origin).
- * We want to ensure the following properties:
- * - All path items specified by the user appear in-order in the OP list.
- * - The first OP in the list is the first path item specified by the user.
- * - The last OP in the list is the last path item specified by the user.
- * - TODO: it doesn't order ops not at origin/destination, find a clean way to handle this case
- */
-export const sortPathOperationalPoints = (
-  ops: CoreOperationalPointOnPath[],
-  path: PathItem[]
-): CoreOperationalPointOnPath[] =>
-  ops.toSorted((a, b) => {
-    if (a.position !== b.position) return a.position - b.position;
-    const aPathIndex = path.findIndex((pathItem) => matchOpRefAndWaypoint(pathItem.location, a));
-    const bPathIndex = path.findIndex((pathItem) => matchOpRefAndWaypoint(pathItem.location, b));
-    const lastIndex = path.length - 1;
-    const aIsOrigin = aPathIndex === 0;
-    const bIsOrigin = bPathIndex === 0;
-    const aIsDestination = aPathIndex === lastIndex;
-    const bIsDestination = bPathIndex === lastIndex;
-    if (aIsOrigin && !bIsOrigin) return -1;
-    if (bIsOrigin && !aIsOrigin) return 1;
-    if (aIsDestination && !bIsDestination) return 1;
-    if (bIsDestination && !aIsDestination) return -1;
-    if (aPathIndex !== -1 && bPathIndex !== -1) return aPathIndex - bPathIndex;
-    return 0;
-  });
 
 /**
  * Prepare data to be used in simulation results
